Move app to application
