+++ app/__main__.py
@@ -12,9 +12,10 @@ def cli():
 @cli.command()
 @click.option("--host", default="0.0.0.0", help="Address to listen on to run on")
 @click.option("--port", default=8000, help="Port to run on")
-def run(host, port):
+@click.option("--reload", is_flag=True, default=False, help="Enable auto-reload.")
+def run(host, port, reload):
     """Run the application."""
-    uvicorn.run("app:app", host=host, port=port)
+    uvicorn.run("app.application:app", host=host, port=port, reload=reload)
 
 
 if __name__ == "__main__":
