+++ app/cli/import-data.py
@@ -8,7 +8,6 @@ from contextlib import closing
 import sqlalchemy
 from tqdm import tqdm
 
-from app import configure_database_session_manager
 from app.cli import curate, utils
 from app.db.model import (
     AnalysisSoftwareSourceCode,
@@ -32,6 +31,7 @@ from app.db.model import (
     SingleCellExperimentalTrace,
     SingleNeuronSimulation,
 )
+from app.db.session import configure_database_session_manager
 
 
 def get_or_create_annotation_body(annotation_body, db):
